Use new catalogueCategories api functions
